mcp-servers/deep-think/requirements.txt
mcp>=1.2.0
google-genai>=0.8.0
